% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtools-package.R
\docType{data}
\name{ELISA}
\alias{ELISA}
\title{Data from an ELISA assay}
\format{
a data frame with the following columns: \itemize{
\itemPlateDayfactor. Specifies one of four physically disctinct 96 well
plates \itemReadfactor. The signal was read 3 times for each plate.
\itemDescriptioncharacter. Indicates contents of sample.
\itemConcentrationnumeric. Nominal concentration of standards (NA for all
other samples).  \itemSignalnumeric. Assay signal. Specifically, optical
density (a colorimetric assay).  }
}
\source{
Anonymized data.
}
\description{
Observed signals and (for some observations) nominal concentrations for
samples that were aliquoted to multiple assay plates, which were read
multiple times on multiple days.
}
\keyword{datasets}
