\name{gtools-deprecated}
\alias{gtools-deprecated}
\alias{addLast}
\title{Deprecated Functions in the gtools package}
\description{
  These functions are provided for compatibility with older versions of
  gtools, and may be defunct as soon as the next release.
}
\usage{
addLast(fun)
}
\arguments{
  \item{fun}{Function to be called.}
}
\details{
  The original help page for these functions is often
  available at \code{help("oldName-deprecated")} (note the quotes).

  \itemize{
    \item{ \code{addLast} has been replaced by \code{lastAdd}, which has
    the same purpose but appled using different syntax. }
    }
    
}
\seealso{
  \code{\link{Deprecated}}
  \code{\link[gtools]{lastAdd}}
}
\keyword{misc}
