% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newVersionAvailable.R
\name{newVersionAvailable}
\alias{newVersionAvailable}
\title{Check if a new version of R is available}
\usage{
newVersionAvailable(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. If TRUE, no}
}
\value{
The \code{numeric_version} of a newer version, if available, otherwise NULL.
}
\description{
Compare the current version of R to the versions available on CRAN and display
the results.
}
\examples{

newVersionAvailable() # Display result
flag <- newVersionAvailable(quiet=TRUE) # Don't display results
str(flag)

}
