% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_color_super}
\alias{scale_color_super}
\title{Scale color with SUPE_R}
\usage{
scale_color_super(
  palette = "cool",
  discrete = TRUE,
  reverse = FALSE,
  n_colors = 16,
  ...
)
}
\arguments{
\item{discrete}{Logical}

\item{reverse}{Reverse palette}

\item{pallette}{Palette to create color ramp from. Options: main, cool, hot, mixed, grey}
}
\description{
Scale color with SUPE_R
}
