% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_path.R
\name{split_path}
\alias{split_path}
\title{Split a File Path into Components}
\usage{
split_path(x, depth_first = TRUE)
}
\arguments{
\item{x}{character scalar.  Path to be processed.}

\item{depth_first}{logical.  Should path be returned depth first?  Defaults
to \code{TRUE}.}
}
\value{
Character vector of path components, depth first.
}
\description{
This function converts a character scalar containing a
 \emph{valid} file path into a character vector of path components
 (e.g. directories).
}
\example{
    longPath = "/users/me/src/this/is/a/long/path"
    split_path(longPath)
}