% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badDend.R
\docType{data}
\name{badDend}
\alias{badDend}
\title{Dataset That Crashes Base:::Plot.Dendogram with 'Node Stack Overflow'}
\format{
The format is: num [1:2047, 1:12] 1 2 3 4 5 6 7 8 9 10 ...  -
attr(*, "dimnames")=List of 2 ..$ : NULL ..$ : chr [1:12] "X" "V1" "V2" "V3"
...
}
\description{
Base:::Plot.Dendogram() will generate a 'Node Stack Overflow' when run on a
dendrogram appropriately constructed from this data set.
}
\note{
See help page for \code{\link{unByteCode}} to see how to construct the
'bad' dentrogram from this data and how to work around the issue.
}
\examples{

data(badDend)
}
\keyword{datasets}
