\name{gtools-defunct}
\alias{gtools-defunct}
\alias{assert}
\title{Deprecated Functions in the gtools package}
\description{
  These functions are no longer available in gtools.
}
\usage{
assert(FLAG)
}
\arguments{
  \item{FLAG}{ Expression that should evaluate to a boolean vector}
}
\details{
  \code{assert} is a defunct synonym for \code{\link[base]{stopifnot}}.
}
\seealso{
  \code{\link{Defunct}}
}
\keyword{misc}
