\name{gtools-deprecated}
\alias{gtools-deprecated}
\title{Deprecated Functions in the gtools package}
\description{
  These functions are provided for compatibility with older versions of
  gtools, and may be defunct as soon as the next release.
}
%\usage{
%}
%\arguments{
%}
\details{
  gtools currently contains no deprecated functions.
  
%  The original help page for these functions is often
%  available at \code{help("oldName-deprecated")} (note the quotes).
%
%  \itemize{
%    \item{}
%    }
    
}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
