% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Randomly Permute the Elements of a Vector}
\usage{
permute(x)
}
\arguments{
\item{x}{Vector of items to be permuted}
}
\value{
Vector with the original items reordered.
}
\description{
Randomly Permute the elements of a vector
}
\details{
This is simply a wrapper function for \code{\link{sample}}.
}
\examples{

  x <- 1:10
  permute(x)

}
\seealso{
\code{\link{sample}}
}
\author{
Gregory R. Warnes \email{greg@warnes.net}
}
\keyword{distribution}
