% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\name{scat}
\alias{scat}
\title{Display debugging text}
\usage{
scat(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{cat}}
}
\value{
NULL (invisibly)
}
\description{
If \code{getOption('DEBUG')==TRUE}, write text to STDOUT and flush so that
the text is immediatly displayed. Otherwise, do nothing.
}
\examples{

options(DEBUG = NULL) # makee sure DEBUG isn't set
scat("Not displayed")

options(DEBUG = TRUE)
scat("This will be displayed immediately (even in R BATCH output \n")
scat("files), provided options()$DEBUG is TRUE.")
}
\seealso{
\code{\link[base]{cat}}
}
\author{
Gregory R. Warnes \email{greg@warnes.net}
}
\keyword{print}
